+++ src/Tools/ApiTests.ts
@@ -14,6 +14,7 @@ import * as BREnvironment from '../BREnvironment';
 import * as BRConfiguration from '../BRConfiguration';
 import * as BRAsProjectWorkspace from '../BRAsProjectWorkspace';
 import * as BrAsProjectFiles from '../BrAsProjectFiles';
+import {Logger, LogLevel} from '../BrLog';
 //import * as NAME from '../BRxxxxxx';
 
 
@@ -63,6 +64,9 @@ async function testCommand(arg1: any, arg2: any, context: vscode.ExtensionContex
 	}
 	if (await Dialogs.yesNoDialog('Run tests for VS Code extension context?')) {
 		await testVsCodeExtensionContext(context);
+	}
+	if (await Dialogs.yesNoDialog('Run tests for BrLog')) {
+		await testBrLog(context);
 	}
 	// end
 	Helpers.logTimedHeader('Test command end');
@@ -454,5 +458,28 @@ async function testVsCodeExtensionContext(context: vscode.ExtensionContext) : Pr
 	context.workspaceState;
 	context.globalState;
 	console.warn('Test VsCodeExtensionContext end');
+}
+
+
+async function testBrLog(context: vscode.ExtensionContext): Promise<void> {
+	console.warn('Test BrLog start');
+	// log messages of various levels
+	Logger.default.fatal("Some fatal 1");
+	Logger.default.error("Some error 1");
+	Logger.default.warning("Some warning 1");
+	Logger.default.info("Some info 1");
+	Logger.default.debug("Some debug 1");
+	// log with objects and array
+	Logger.default.fatal("Now log with additional data!");
+	Logger.default.fatal("Undefined", { data: undefined });
+	Logger.default.fatal("Null", { data: null });
+	Logger.default.fatal("Boolean", { data: false });
+	Logger.default.fatal("Number", { data: 33 });
+	Logger.default.fatal("String", { data: "hello" });
+	const someObj = { a: "hello", b: "world", c: 33, d: { d1: 20, d2: "42" } };
+	Logger.default.fatal("Obj", { data: someObj });
+	const someArray = [{ a: 33, b: { b1: 33, b2: false } }, undefined, "hello", 33, false, null, { q: "testQ", r: "testR" }];
+	Logger.default.fatal("Array", { data: someArray });
+	console.warn('Test BrLog end');
 }